Add leather belt

+++ code/Data/StashableDetailData.cs
@@ -7,6 +7,6 @@ internal class StashableDetailData
 	public string Identity { get; set; }
 	public int StashSlot { get; set; }
 	public int Quantity { get; set; }
-	public byte Durability { get; set; }
+	public int Durability { get; set; }
 
 }
